new merge

+++ CodeLite/TerminalEmulator.cpp
@@ -2,6 +2,9 @@
 #include <wx/filename.h>
 #include <wx/log.h>
 #include "processreaderthread.h"
+#include "macros.h"
+#include <algorithm>
+#include "dirsaver.h"
 
 #ifndef __WXMSW__
 #include <signal.h>
@@ -10,28 +13,63 @@
 wxDEFINE_EVENT(wxEVT_TERMINAL_COMMAND_EXIT, clCommandEvent);
 wxDEFINE_EVENT(wxEVT_TERMINAL_COMMAND_OUTPUT, clCommandEvent);
 
-BEGIN_EVENT_TABLE(TerminalEmulator, wxEvtHandler)
-EVT_COMMAND(wxID_ANY, wxEVT_PROC_DATA_READ, TerminalEmulator::OnProcessOutput)
-EVT_COMMAND(wxID_ANY, wxEVT_PROC_TERMINATED, TerminalEmulator::OnProcessTerminated)
-END_EVENT_TABLE()
+class MyProcess : public wxProcess
+{
+public:
+    TerminalEmulator* m_parent;
+
+public:
+    MyProcess(TerminalEmulator* parent)
+        : wxProcess(parent)
+        , m_parent(parent)
+    {
+        if(m_parent) {
+            m_parent->m_myProcesses.push_back(this);
+        }
+    }
+
+    virtual ~MyProcess() { m_parent = NULL; }
+    void OnTerminate(int pid, int status)
+    {
+        if(m_parent) {
+            clCommandEvent terminateEvent(wxEVT_TERMINAL_COMMAND_EXIT);
+            m_parent->AddPendingEvent(terminateEvent);
+            m_parent->m_pid = wxNOT_FOUND;
+
+            std::list<wxProcess*>::iterator iter = std::find_if(m_parent->m_myProcesses.begin(),
+                                                                m_parent->m_myProcesses.end(),
+                                                                [&](wxProcess* proc) { return proc == this; });
+            if(iter != m_parent->m_myProcesses.end()) {
+                m_parent->m_myProcesses.erase(iter);
+            }
+            delete this;
+        }
+    }
+};
 
 TerminalEmulator::TerminalEmulator()
     : m_process(NULL)
+    , m_pid(wxNOT_FOUND)
 {
+    Bind(wxEVT_ASYNC_PROCESS_OUTPUT, &TerminalEmulator::OnProcessOutput, this);
+    Bind(wxEVT_ASYNC_PROCESS_TERMINATED, &TerminalEmulator::OnProcessTerminated, this);
 }
 
-TerminalEmulator::~TerminalEmulator() {}
+TerminalEmulator::~TerminalEmulator()
+{
+    Unbind(wxEVT_ASYNC_PROCESS_OUTPUT, &TerminalEmulator::OnProcessOutput, this);
+    Unbind(wxEVT_ASYNC_PROCESS_TERMINATED, &TerminalEmulator::OnProcessTerminated, this);
+    std::for_each(m_myProcesses.begin(), m_myProcesses.end(), [&](wxProcess* proc) {
+        MyProcess* myproc = dynamic_cast<MyProcess*>(proc);
+        myproc->m_parent = NULL;
+    });
+}
 
 bool TerminalEmulator::ExecuteConsole(const wxString& command,
                                       const wxString& workingDirectory,
                                       bool waitOnExit,
                                       const wxString& title)
 {
-    if(m_process) {
-        // another process is running
-        return false;
-    }
-
     wxString consoleCommand;
     wxString strTitle = title;
     if(strTitle.IsEmpty()) {
@@ -39,18 +77,23 @@ bool TerminalEmulator::ExecuteConsole(const wxString& command,
     } else {
         strTitle.Prepend("'").Append("'");
     }
-
 #ifdef __WXMSW__
     consoleCommand = PrepareCommand(command, strTitle, waitOnExit);
 
 #elif defined(__WXGTK__)
     // Test for the common terminals on Linux
-    if(wxFileName::Exists("/usr/bin/gnome-terminal")) {
+    // gnome-terminal, konsole and lxterminal are all starting asychronously
+    // this means that "waitOnExit" has no effect here
+    if(wxFileName::Exists("/usr/bin/gnome-terminal") && !waitOnExit) {
         consoleCommand << "/usr/bin/gnome-terminal -t " << strTitle << " -x "
                        << PrepareCommand(command, strTitle, waitOnExit);
 
-    } else if(wxFileName::Exists("/usr/bin/konsole")) {
+    } else if(wxFileName::Exists("/usr/bin/konsole") && !waitOnExit) {
         consoleCommand << "/usr/bin/konsole -e " << PrepareCommand(command, strTitle, waitOnExit);
+
+    } else if(wxFileName::Exists("/usr/bin/lxterminal") && !waitOnExit) {
+        consoleCommand << "/usr/bin/lxterminal -T " << strTitle << " -e "
+                       << PrepareCommand(command, strTitle, waitOnExit);
 
     } else if(wxFileName::Exists("/usr/bin/uxterm")) {
         consoleCommand << "/usr/bin/uxterm -T " << strTitle << " -e " << PrepareCommand(command, strTitle, waitOnExit);
@@ -61,12 +104,26 @@ bool TerminalEmulator::ExecuteConsole(const wxString& command,
 
 #elif defined(__WXMAC__)
 
+    consoleCommand = TERMINAL_CMD;
+    consoleCommand.Replace("$(CMD)", command);
+    wxUnusedVar(strTitle);
+    wxUnusedVar(waitOnExit);
+
 #endif
     if(consoleCommand.IsEmpty()) return false;
     wxLogMessage(consoleCommand);
 
-    m_process = ::CreateAsyncProcess(this, consoleCommand, IProcessCreateConsole, workingDirectory);
-    return m_process != NULL;
+    // Create the process as group leader, this way we make sure that killing it
+    // will also kill all the children processes
+    DirSaver ds;
+    if(!workingDirectory.IsEmpty()) {
+        wxSetWorkingDirectory(workingDirectory);
+        wxLogMessage("Working directory is now: " + ::wxGetCwd());
+    }
+    
+    wxLogMessage(consoleCommand);
+    m_pid = ::wxExecute(consoleCommand, wxEXEC_ASYNC | wxEXEC_MAKE_GROUP_LEADER, new MyProcess(this));
+    return (m_pid != 0);
 }
 
 wxString TerminalEmulator::PrepareCommand(const wxString& str, const wxString& title, bool waitOnExit)
@@ -79,16 +136,16 @@ wxString TerminalEmulator::PrepareCommand(const wxString& str, const wxString& t
     escapedString.Replace("\"", "\\\"");
     command << "/bin/bash -c \"" << escapedString;
     if(waitOnExit) {
-        command << " && echo 'Hit ENTER to continue' && read";
+        command << " ; echo 'Hit ENTER to continue' ; read";
     }
     command << "\"";
 
 #elif defined(__WXMSW__)
     // Windows
     wxString escapedString = str;
-    command << "cmd /c call title " << title << " && " << escapedString;
+    command << "cmd /C call title \"" << title << "\" && " << escapedString;
     if(waitOnExit) {
-        command << " && echo \"\" && pause";
+        command << " && echo \"\" & pause";
     }
 #else
 // OSX
@@ -97,11 +154,11 @@ wxString TerminalEmulator::PrepareCommand(const wxString& str, const wxString& t
     return command;
 }
 
-void TerminalEmulator::OnProcessTerminated(wxCommandEvent& event)
+void TerminalEmulator::OnProcessTerminated(clProcessEvent& event)
 {
     // Process terminated
     wxDELETE(m_process);
-
+    m_pid = wxNOT_FOUND;
     // Notify that the terminal has terminated
     clCommandEvent terminateEvent(wxEVT_TERMINAL_COMMAND_EXIT);
     AddPendingEvent(terminateEvent);
@@ -110,18 +167,22 @@ void TerminalEmulator::OnProcessTerminated(wxCommandEvent& event)
 void TerminalEmulator::Terminate()
 {
     if(IsRunning()) {
-        m_process->Terminate();
+        if(m_process) {
+            m_process->Terminate();
+        }
+        if(m_pid != wxNOT_FOUND) {
+            wxKill(m_pid, wxSIGKILL, NULL, wxKILL_CHILDREN);
+            m_pid = wxNOT_FOUND;
+        }
     }
 }
 
-bool TerminalEmulator::IsRunning() const { return m_process != NULL; }
+bool TerminalEmulator::IsRunning() const { return (m_process != NULL) || (m_pid != wxNOT_FOUND); }
 
-void TerminalEmulator::OnProcessOutput(wxCommandEvent& event)
+void TerminalEmulator::OnProcessOutput(clProcessEvent& event)
 {
-    ProcessEventData* ped = reinterpret_cast<ProcessEventData*>(event.GetClientData());
     clCommandEvent evtOutput(wxEVT_TERMINAL_COMMAND_OUTPUT);
-    evtOutput.SetString(ped->GetData());
-    wxDELETE(ped);
+    evtOutput.SetString(event.GetOutput());
     AddPendingEvent(evtOutput);
 }
